fixing the flag of exec resource

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,4 +1,5 @@
 # This resource is to exec a command 
 exec {'killmenow':
 command => 'pkill -f killmenow',
+onlyif  => 'running'
 }
